Rename MeasureAsDist to MeasureAsDistArgument
